Fix the previous problem more succinctly
It was actually just that for false it *had* to be a bool, and for the
true case it could be just about anything (anything 'truthy').

This is the right fix, I was totally wrong. My mistake. I had type
unsafety on my mind and unfairly threw synapse under a bus :(